refactor : extract icon_content and reusable_card
