+++ lib/bmicalculator/icon_content.dart
@@ -1,0 +1,36 @@
+import 'package:flutter/material.dart';
+
+const labelTextStyle = TextStyle(
+  fontSize: 18,
+  color: Color(0xFF8D8E98),
+);
+
+class IconContent extends StatelessWidget {
+  const IconContent({
+    required this.icon,
+    required this.name
+  });
+
+  final IconData icon;
+  final String name;
+
+  @override
+  Widget build(BuildContext context) {
+    return Column(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: [
+        Icon(
+          icon,
+          size: 80,
+        ),
+        const SizedBox(
+          height: 15,
+        ),
+        Text(
+          name,
+          style: labelTextStyle
+        )
+      ],
+    );
+  }
+}
