+++ lib/bmicalculator/reusable_card.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class ReusableCard extends StatelessWidget {
+  const ReusableCard({required this.color, required this.cardChild});
+
+  final Color color;
+  final Widget cardChild;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.all(10),
+      decoration: BoxDecoration(
+        color: const Color(0xFF1D1E33),
+        borderRadius: BorderRadius.circular(10),
+      ),
+      child: cardChild,
+    );
+  }
+}
